Progress: split into one class per file
* 3 classes / files: IProgress, ConsoleProgress, NoDisplayProgress.
* Add a cpp file just for ConsoleProgress (the implementation is more
'complicated' than the others).
