+++ src/AvTranscoder/File/InputFile.hpp
@@ -10,7 +10,7 @@
 #include <AvTranscoder/CodedStream/AvInputStream.hpp>
 
 #include <AvTranscoder/Metadatas/MediaMetadatasStructures.hpp>
-#include <AvTranscoder/progress/ProgressListener.hpp>
+#include <AvTranscoder/progress/IProgress.hpp>
 
 #include <AvTranscoder/Profile.hpp>
