+++ src/AvTranscoder/avTranscoder.i
@@ -50,7 +50,7 @@
 #include <AvTranscoder/Transcoder/Transcoder.hpp>
 %}
 
-%include "AvTranscoder/progress/ProgressListener.i"
+%include "AvTranscoder/progress/progress.i"
 
 namespace std {
 %template(IntPair)         pair< size_t, size_t >;
