+++ src/AvTranscoder/progress/ConsoleProgress.cpp
@@ -1,4 +1,4 @@
-#include "ProgressListener.hpp"
+#include "ConsoleProgress.hpp"
 
 #include <iostream>
 #include <iomanip>
@@ -6,11 +6,9 @@
 namespace avtranscoder
 {
 
-IProgress::~IProgress()
-{}
-
 ConsoleProgress::~ConsoleProgress()
-{}
+{
+}
 
 EJobStatus ConsoleProgress::progress( const double processedDuration, const double programDuration )
 {
@@ -23,14 +21,6 @@ EJobStatus ConsoleProgress::progress( const double processedDuration, const doub
 	// if( processedFrames >= 100 )
 	// 	return avtranscoder::eJobStatusCancel;
 
-	return eJobStatusContinue;
-}
-
-NoDisplayProgress::~NoDisplayProgress()
-{}
-
-EJobStatus NoDisplayProgress::progress( const double processedDuration, const double programDuration )
-{
 	return eJobStatusContinue;
 }
