+++ src/AvTranscoder/progress/progress.i
@@ -1,0 +1,15 @@
+%module(directors="1") AvTranscoder
+
+
+%{
+ #include <AvTranscoder/progress/IProgress.hpp>
+ #include <AvTranscoder/progress/ConsoleProgress.hpp>
+ #include <AvTranscoder/progress/NoDisplayProgress.hpp>
+%}
+
+/* turn on director wrapping for IProgress */
+%feature("director") IProgress;
+
+%include <AvTranscoder/progress/IProgress.hpp>
+%include <AvTranscoder/progress/ConsoleProgress.hpp>
+%include <AvTranscoder/progress/NoDisplayProgress.hpp>
